+++ .eslintrc
@@ -2,6 +2,7 @@
   "env": {
     "es6": true,
     "browser": true,
+    "commonjs": true
   },
   "root": true,
   "plugins": [],
