+++ .gitignore
@@ -45,3 +45,7 @@ jspm_packages
 
 # Yarn Integrity file
 .yarn-integrity
+
+.DS_Store
+
+scripts/bundle.js
