+++ index.html
@@ -3,7 +3,7 @@
 <html lang="en">
 	<head>
 		<meta charset="utf-8" />
-		<script src="./scripts/fluid.js"></script>
+		<script src="./scripts/bundle.js"></script>
 	</head>
 
 	<body>
